+++ vacation-planner-pro/src/main/frontend/src/__test__/LogIn.test.js
@@ -1,9 +1,14 @@
 import { render } from '@testing-library/react';
 import LogIn from '../pages/LogIn';
+import { BrowserRouter, Router } from 'react-router-dom';
 
 describe('LogIn', () => {
   test('That company logo displays on log in page', () => {
-    const { getByAltText } = render(<LogIn />);
+    const { getByAltText } = render(
+      <BrowserRouter>
+        <LogIn />
+      </BrowserRouter>,
+    );
     const image = getByAltText('Company Logo');
 
     expect(image).toBeInTheDocument();
